+++ src/game/server/entities/point/effects.h
@@ -72,7 +72,7 @@ public:
 	}
 	inline void SetTexture(int spriteIndex) { pev->modelindex = spriteIndex; }
 	inline void SetScale(float scale) { pev->scale = scale; }
-	inline void SetColor(int r, int g, int b) { SetRenderColor({r, g, b}); }
+	inline void SetColor(int r, int g, int b) { SetRenderColor({static_cast<float>(r), static_cast<float>(g), static_cast<float>(b)}); }
 	inline void SetBrightness(int brightness) { pev->renderamt = brightness; }
 
 	inline void AnimateAndDie(float framerate)
@@ -126,7 +126,7 @@ public:
 	inline void SetTexture(int spriteIndex) { pev->modelindex = spriteIndex; }
 	inline void SetWidth(int width) { pev->scale = width; }
 	inline void SetNoise(int amplitude) { pev->body = amplitude; }
-	inline void SetColor(int r, int g, int b) { SetRenderColor({r, g, b}); }
+	inline void SetColor(int r, int g, int b) { SetRenderColor({static_cast<float>(r), static_cast<float>(g), static_cast<float>(b)}); }
 	inline void SetBrightness(int brightness) { pev->renderamt = brightness; }
 	inline void SetFrame(float frame) { pev->frame = frame; }
 	inline void SetScrollRate(int speed) { pev->animtime = speed; }
@@ -151,7 +151,7 @@ public:
 	inline int  GetTexture() { return pev->modelindex; }
 	inline int  GetWidth() { return pev->scale; }
 	inline int  GetNoise() { return pev->body; }
-	// inline void GetColor( int r, int g, int b ) { SetRenderColor({r, g, b}); }
+	// inline void GetColor( int r, int g, int b ) { SetRenderColor({static_cast<float>(r), static_cast<float>(g), static_cast<float>(b)}); }
 	inline int  GetBrightness() { return pev->renderamt; }
 	inline int  GetFrame() { return pev->frame; }
 	inline int  GetScrollRate() { return pev->animtime; }
